Added tests to intx, fixed lots of bugs in intx and added fast multiplication.

+++ code/mathematics/fft.cpp
@@ -1,5 +1,5 @@
 #include <complex>
-typedef complex<double> cpx;
+typedef complex<long double> cpx;
 void fft(cpx *x, int n, bool inv=false) {
     for (int i = 0, j = 0; i < n; i++) {
         if (i < j) swap(x[i], x[j]);
